Map services to standard https port
Follow the Bright KB config to provide access to cluster services
via standard port 443 by setting up proxy configs to the
default port maps for bright-view, userportal, and openstack services.

http://kb.brightcomputing.com/faq/index.php?action=artikel&cat=9&id=291

The KB addresses bright-view and userportal via a rewrite rule
to move requests to for the service to https transport and then
maps https requests to a proxy config to the backend.

This code adopts the same model for openstack proxying.