Fix proxypass configuration for openstack to stay on https
Add a replacement rule for the service directory page to switch
openstack urls to https.

Add https backend for proxypass config to keep the proxy header
re-writes in the https protocol.  It seems that if the backend is http
the the proxypass-ed redirects will retain an http url from the
backend instead of being fixed to be https.  Using an https
backend corrects this unwanted behavior.