Fix too many open files error in openstack dashboard
The dashboard has a file desriptor leak when connecting to keystone
for authentication.  This eventually exceeds the softmax file handles
of 1024 and causes the openstack-dashboard process to stop responding
leading the an nginx bad gateway error when viewing the dashboard.

The bug is documented here https://bugs.launchpad.net/horizon/+bug/1780164
and has a comment with a patch.

This playbook applies the patch